Updated models to store SystemID as a long (to cover 4b+ systems)

+++ taxee_server/taxee_server/Models/Body.cs
@@ -4,6 +4,6 @@
     {
         public string BodyID { get; set; }
         public string BodyName { get; set; }
-        public int SystemID { get; set; }
+        public long SystemID { get; set; }
     }
 }
